Prototype for auto-converting some structs to hashes
Basically, this is a bunch of scripts that do an immensely stupid scan
of the headers for a certain struct, then do a naive parse thereof, then
extract the types, read all typemaps (distribution and core perl), then
expand those typemaps in a rather limited way, but generally akin to
how ExtUtils::ParseXS does it, then use that and some extra code
generation to create functions that convert such a C struct to a Perl
hash.

The same approach is unlikely to work in the reverse direction since the
code so far skips members with types it doesn't support and even moreso
because of type/completeness/etc checking.